LayoutTests/fast/dom/Element/id-in-param.html

<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css">
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
    
<object width=100 height=100>
  <param name="movie" id="id1">
</object>    

<script>
description("Test using id in param element.");

shouldBeNonNull('document.getElementById("id1")');

var successfullyParsed = true;
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
