third_party/blink/web_tests/platform/generic/fast/dom/Element/id-in-param-expected.txt

Test using id in param element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("id1") is non-null.
PASS successfullyParsed is true

TEST COMPLETE
